% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carmon-package.R
\docType{package}
\name{carmon-package}
\alias{carmon}
\alias{carmon-package}
\title{carmon: Copula Aided Reconstruction of Multi-Omics Networks}
\description{
.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrQuestion/carmon}
  \item Report bugs at \url{https://github.com/DrQuestion/carmon/issues}
}

}
\author{
\strong{Maintainer}: Alessio Albanese \email{alessio.albanese@wur.nl} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
