% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carmon.R
\name{carmon}
\alias{carmon}
\title{Perform copula-aided reconstruction, analysis and plot of a multi-omics network}
\usage{
carmon(
  layers,
  ...,
  p = NULL,
  omics = NULL,
  marginals = NULL,
  noninv_method = NULL,
  copula = NULL,
  net_method = "coglasso",
  sel_method = NULL,
  cor_cutoff = NULL,
  cor_quant = NULL,
  minimal_output = FALSE,
  analyse = TRUE,
  c_measures = "dbce",
  max_candidates_c_measures = NULL,
  quantile_c_measures = NULL,
  scaled_c_measures = TRUE,
  plot = TRUE,
  plot_node_labels = TRUE,
  hide_isolated = TRUE,
  plot_hot_nodes = TRUE,
  verbose = 1
)
}
\arguments{
\item{layers}{The omics layers to analyze. Preferably provided as a named R
list of \strong{non-normalized} omics data sets. If possible the names of the list
should correspond to the respective omics type. If not possible, for
example because of two layers from the same technology, please provide the
omics types with the parameter \code{omics}.  To see a list of available
omics types use the function \code{which_omics()}.\cr
Each data set should be source-matched (same amount of matched samples or
individuals across each data set). Placing of the samples (or individuals)
should also be consistent: either along the rows for \emph{all} the data
sets, or along the columns for \emph{all} the data sets, nothing in
between. All the samples (or individuals) should also have consistent
naming across the data sets.\cr
\code{layers} can also be a single unified data set, but then it is necessary to
specify the argument \code{p}.}

\item{...}{The additional optional arguments to be given for the network
reconstruction and selection procedure. The available arguments depend on
the chosen network reconstruction and network selection methods. If using
\code{"coglasso"} (default option), see \code{\link[coglasso:bs]{coglasso::bs()}} for both reconstruction
and selection arguments. If reconstructing with \code{"glasso"}, \code{"mb"}, or
\code{"correlation"} without setting \code{cor_cutoff} or \code{cor_quant} see,
respectively, \code{\link[huge:huge.glasso]{huge::huge.glasso()}}, \code{\link[huge:huge.mb]{huge::huge.mb()}}, or
\code{\link[huge:huge.ct]{huge::huge.ct()}}. For additional options for network selection from these
last three reconstruction methods, see \code{\link[huge:huge.select]{huge::huge.select()}}.}

\item{p}{Necessary \strong{only} in case \code{layers} is a single data set.
A vector with with the number of variables for each omic layer of the
data set (e.g. the number of transcripts, metabolites etc.), in the same
order the layers have in the data set. If the sum of the elements in \code{p}
is less than the real number of features in the data set, \code{carmon()}
assumes that there is an omics layer more, with as many features as the
real number of features minus \code{sum(p)}.}

\item{omics}{Highly recommended. A character vector of as many elements as
the number of omics layers, naming which omics types each layer contains,
in the same order as provided in the input \code{layers}
(\emph{e.g.} \code{omics = c("RNA-seq", "proteomics", "metabolomics")}). To
see a list of terms and omics technologies for which \pkg{carmon} is
specifically tailored, use the function \code{which_omics()}.}

\item{marginals}{Optional, to be specified when the user prefers to use
different marginal distributions than the default distribution \pkg{carmon}
tailored for each omics layers. A vector of as many elements as the number
of layers, specifying which marginal distribution should be used for each
omics layer, in the same order as provided in the input \code{layers}. To see a
list of available marginal distributions, use the function
\code{which_marginals()}. For a mixed custom setting, place a \code{0} in the vector
in the position corresponding to the omics layers for which the default
distribution is desired. Otherwise, specify the desired marginal
distribution (\emph{e.g.} \code{marginals = c(0, "lognormal", 0)} means
"default, lognormal, default").}

\item{noninv_method}{A placeholder for future functionalities of
\pkg{carmon}, do not use.}

\item{copula}{A placeholder for future functionalities of \pkg{carmon}, do
not use.}

\item{net_method}{The network reconstruction method to use. The four methods
currently available are: \code{"coglasso"} for \emph{collaborative graphical lasso},
\code{"glasso"} for \emph{graphical lasso}, \code{"mb"} for the Meinshausen-Buhlmann
neighborhood selection, and \code{"correlation"} for a thresholded Pearson's
correlation network. The default method is \code{"coglasso"}. See References for
more information on the network reconstruction methods.}

\item{sel_method}{The network selection method. Each reconstruction procedure
has its own set of model selection procedures available. See References for
more information on the network selection methods. Here is the list of
them:
\itemize{
\item for \code{net_method = "coglasso"}: \code{"xstars"} for \emph{eXtended StARS},
\code{"xestars"} for \emph{eXtended  Efficient StARS}, and \code{"ebic"} for Extended
Bayesian Information Criterion. Default is \code{"xestars"}.
\item for \code{net_method = "glasso"}: \code{"stars"} for \emph{StARS}, \code{"ric"} for
Rotational Information Criterion, and \code{"ebic"} for Extended Bayesian
Information Criterion. Default is \code{"stars"}.
\item for \code{net_method = "mb"} or \code{"correlation"}: \code{"stars"} for \emph{StARS},
and \code{"ric"} for Rotational Information Criterion. Default is \code{"stars"}.
}}

\item{cor_cutoff}{Optional for \code{net_method = "correlation"}. The cutoff
value for the absolute Pearson's correlation network. Any edge with an
absolute correlation below the cutoff is excluded from the final network.
Not used by default, as the cutoff is generated by internal calculations.}

\item{cor_quant}{Optional for \code{net_method = "correlation"}, to set as an
alternative to \code{cor_cutoff}. Determine the cutoff of correlation based on
the top percentile indicated by the user. For example, \code{cor_quant = 0.2}
would set as a correlation cutoff the 20th percentile of the absolute
Pearson's correlation values, ordered from highest to lowest. Not used by
default, as the cutoff is generated by internal calculations.}

\item{minimal_output}{Logical. Set to \code{TRUE} to get only a minimal output from
the network reconstruction and selection module of \code{carmon()}, which mainly
differs on the basis of on the network reconstruction method set with
\code{net_method.} Defaults to \code{FALSE}.}

\item{analyse}{Logical. Whether to perform the consensus centrality analysis
to identify key omics features. The highlighted features are important
based on a consensus of multiple centrality measures. Defaults to \code{TRUE}.}

\item{c_measures}{A string of characters, each one representing one of four
possible centrality measures implemented in the consensus centrality
analysis: \code{"d"} stands for degree centrality, \code{"b"} for betweenness
centrality, \code{"c"} for closeness centrality, and \code{"e"} for eigenvector
centrality. Default is \code{"dbce"}: all the four measures.}

\item{max_candidates_c_measures}{What is the highest amount of nodes that can
be highlighted by each measure of centrality, \emph{before} finding the
consensus? Default is 20. When given together with the \code{quantile_c_measures}
argument, it overrides the \code{quantile_c_measures} argument when
\code{max_candidates_c_measures} is smaller than  the chosen quantile, it is
overrode by \code{quantile_c_measures} in the opposite case.}

\item{quantile_c_measures}{What is the top percentile of nodes that can be
highlighted by each measure of centrality, \emph{before} finding the
consensus? Default is the top 5\%. When given together with the
\code{max_candidates_c_measures}, it overrides the \code{max_candidates_c_measures}
argument when the amount of nodes in the chosen top percentile is smaller
than the chosen maximum amount of candidate nodes, it is overrode by
\code{max_candidates_c_measures} in the opposite case.}

\item{scaled_c_measures}{Logical. Whether to compute centrality measures as
0-1 scaled values. Defaults to \code{TRUE}.}

\item{plot}{Logical. Whether to plot the multi-omics network (enriched by the
results of the centrality analysis when that is performed), and the results
of the centrality analysis (when performed). Defaults to \code{TRUE}.}

\item{plot_node_labels}{Show node names in the plot of the network. Defaults
to \code{TRUE}.}

\item{hide_isolated}{Hide from the plot nodes that are not connected to any
other node. Defaults to \code{TRUE}.}

\item{plot_hot_nodes}{Highlight in the plot nodes that are selected as
consensus-central nodes by the node centrality analysis, by drawing them
bigger and adding the criteria for which they result to be central. The
larger the consensus, the bigger the nodes are plotted. Defaults to \code{TRUE}.}

\item{verbose}{The level of verbosity of the \code{carmon()} process. \code{0}
suppresses the information output, while \code{1} and \code{2} give progressively
increasing amounts of information about the inner computations happening
inside the package's modules. Defaults to \code{1}}
}
\value{
\code{carmon()} returns an object of \code{S3} class \code{carmon}. The elements of
this object depend on the chosen network reconstruction (\code{"coglasso"}, by
default) and selection (\code{"xestars"}, by default) strategies, and on whether
or not the consensus centrality analysis is carried out (yes, by default).
\cr Every possible running mode of \code{carmon()} produces a \code{carmon} object
with the following elements:
\itemize{
\item \code{layers} is an R list, each element being a data set of the
corresponding omics layer, already copulized and transferred to the
normal realm.
\item \code{omics} is a vector containing the omics type assigned to each omics
layer.
\item \code{marginals} is a vector containing the marginal distributions used to
transfer each omics layer to the normal realm.
\item \code{sel_adj} is the adjacency matrix of the final selected \pkg{carmon}
network.
\item \code{net_method} is the chosen network reconstruction method.
\item \code{sel_method} is the chosen model selection method.
\item \code{network} is the \code{igraph} network object of the final selected
\pkg{carmon} network.
\item \code{call} is the matched call.
}

Depending on the different network reconstruction and selection strategies
chosen, some of the other elements may change. When choosing \code{"coglasso"} or
\code{"glasso"} as network reconstruction method, an additional returned element
is:
\itemize{
\item \code{sel_icov} is the inverse covariance matrix of the final selected
network.
}
Otherwise, if choosing "correlation", \code{carmon()} also returns:
\itemize{
\item \code{cor} is the Pearson's correlation matrix of the final selected
network.
}
When \code{minimal_output} is set to \code{TRUE}, these are also the only elements
returned in the object of \code{S3} class \code{carmon} given by \code{carmon()}. For
the non minimal output elements resulting from network reconstruction and
selection, please look at \code{\link[coglasso:bs]{coglasso::bs()}} for \code{net_method = "coglasso"}.
When using \code{"glasso"}, \code{"mb"}, and when using \code{"correlation"} without setting
any value for \code{cor_cutoff} or \code{cor_quant}, look at \code{\link[huge:huge]{huge::huge()}} and at
\code{\link[huge:huge.select]{huge::huge.select()}}.\cr

Moreover, when \code{analyse = TRUE} (default behaviour), the \code{carmon} object
contains two additional elements resulting from the consensus centrality
analysis:
\itemize{
\item \code{report} an R data frame. The rows correspond to the nodes identified to
be central by the analysis, and they are ordered based on how large is
the consensus among the different measures. The data frame has 6 columns: \cr
\emph{candidate}, the name of the central node; \cr
\emph{degree}, the degree centrality of the node; \cr
\emph{betweenness}, the betweenness centrality of the node; \cr
\emph{closeness}, the closeness centrality of the node; \cr
\emph{eigenvector}, the eigenvector centrality of the node; and \cr
\emph{central for}, the a string reporting the first letter of all the
measures according to which the node is central for.
\item \code{measures_list} an R named list of as many elements as the number of
chosen centrality measures, the name of each element being the associated
centrality measure. Each element is a named numerical vector, containing
the measures of the top central nodes identified in the analysis. The
name of each element of the vectors is the name of the node associated to
the reported measure.
}
}
\description{
Single wrapper function encompassing all the functional units of the \pkg{carmon}
package: copula-based transition of non-normal data to the normal realm,
network reconstruction and selection, consensus centrality analysis of the
network to identify key omics features, and enriched plot of the network and
of the analysis results.
}
\examples{
# Suggested usage: provide input data as a named R list, each element being
# the data set of an omics layer, then let carmon do the rest. See the
# vignettes for a more custom usage!
c_obj <- carmon(multi_omics_small, verbose = FALSE)

}
\references{
For \emph{collaborative graphical lasso} and \emph{eXtended StARS}, see
\href{https://doi.org/10.48550/arXiv.2403.18602}{Albanese \emph{et al.} (2024)}

For \emph{graphical lasso}, see
\href{https://doi.org/10.1093/biostatistics/kxm045}{Friedman \emph{et al.} (2007)}

For neighborhood selection, see
\href{https://doi.org/10.1214/009053606000000281}{Meinshausen and Buhlmann (2006)}

For the \emph{StARS} selection procedure, see
\href{https://proceedings.neurips.cc/paper_files/paper/2010/file/301ad0e3bd5cb1627a2044908a42fdc2-Paper.pdf}{Liu \emph{et al.} (2010)}
}
