% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{centrality_report}
\alias{centrality_report}
\title{Display the table of results of the consensus centrality analysis of the \code{carmon} network}
\usage{
centrality_report(carmon_obj)
}
\arguments{
\item{carmon_obj}{The object of \code{S3} class \code{carmon}, possibly already
resulting from a \emph{full} \code{carmon()} run, or at least undergone
centrality analysis through \code{compute_centrality()}.}
}
\value{
\code{centrality_report()} returns the centrality report of the object of
\code{S3} class \code{carmon} given as an input, in the form of an R data frame.
The rows correspond to the nodes identified to be central by the analysis,
and they are ordered based on how large is the consensus among the
different measures. The data frame has 6 columns: \cr
\emph{candidate}, the name of the central node; \cr
\emph{degree}, the degree centrality of the node; \cr
\emph{betweenness}, the betweenness centrality of the node; \cr
\emph{closeness}, the closeness centrality of the node; \cr
\emph{eigenvector}, the eigenvector centrality of the node; and \cr
\emph{central for}, the a string reporting the first letter of all the
measures according to which the node is central for.
}
\description{
Display the table of results of the consensus centrality analysis of the \code{carmon} network
}
\examples{
# Let's build and analyse a carmon network:
c_obj <- carmon(multi_omics_small, net_method = "correlation",
                cor_quant = 0.25, analysis = TRUE, plot = FALSE,
                # analysis is already TRUE by default
                verbose = FALSE)
# To display the table of the results of the centrality analysis:
centrality_report(c_obj)

}
