% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulize.R
\name{copulize}
\alias{copulize}
\title{Use copulae to transfer every omics layer to the normal realm}
\usage{
copulize(
  layers,
  p = NULL,
  omics = NULL,
  marginals = NULL,
  noninv_method = NULL,
  copula = NULL,
  verbose = FALSE
)
}
\arguments{
\item{layers}{The omics layers to analyze. Preferably provided as a named R
list of \strong{non-normalized} omics data sets. If possible the names of the list
should correspond to the respective omics type. If not possible, for
example because of two layers from the same technology, please provide the
omics types with the parameter \code{omics}.  To see a list of available
omics types run the function \code{which_omics()}.
Each data set should be source-matched (same amount of matched samples or
individuals across each data set). Placing the samples (or individuals)
should also be consistent: either along the rows for \emph{all} the data
sets, or along the columns for \emph{all} the data sets, nothing in
between. All the samples (or individuals) should also have consistent
naming across the data sets.\cr
\code{layers} can also be a single unified data set, but then it is necessary to
specify the argument \code{p}.}

\item{p}{Optional, to be specified only in case layers is a single data set.
A vector with with the number of variables for each omic layer of the
data set (e.g. the number of transcripts, metabolites etc.), in the same
order the layers have in the data set. If given a single number, carmon
assumes that the total of data sets is two, and that the number given is
the dimension of the first one.}

\item{omics}{Highly recommended. A vector of as many elements as the number
of layers, naming what omics each layer contains, in the same order as
provided in the input \code{layers}
(\emph{e.g.} \code{omics = c("RNA-seq", "proteomics", "metabolomics")}). To
see a list of terms and omics technologies for which carmon is specifically
tailored, use the function \code{which_omics()}.}

\item{marginals}{Optional, to be specified when the user prefers to use
different marginal distributions than the default distribution carmon
tailored for each omics layers. A vector of as many elements as the number
of layers, specifying which marginal distribution should be used for each
omics layer, in the same order as provided in the input \code{layers}. To see a
list of available marginal distributions, use the function
\code{which_marginals()}. For a more custom setting, place a \code{0} in the vector
in the position corresponding to the omics layers for which the default
distribution is desired. Otherwise, specify the desired marginal
distribution.}

\item{noninv_method}{A placeholder for future functionalities of carmon, do
not use.}

\item{copula}{A placeholder for future functionalities of carmon, do not use.}

\item{verbose}{The level of verbosity of the copulization process. \code{0}
suppresses the information output, while \code{1} and \code{2} give progressively
increasing amounts of information about the inner computations happening
inside \code{copulize()}.}
}
\value{
\code{copulize()} returns an object of \code{S3} class \code{carmon_cop}, having
the following elements:
\itemize{
\item \code{layers} is an R list, each element being a data set of the corresponding
layer, already copulized and transferred to the normal realm.
\item \code{omics} is a vector containing the omics type assigned to each omics layer.
\item \code{marginals} is a vector containing the marginal distributions used to
transfer each omics layer to the normal realm.
\item \code{copulize_call} is the matched call.
}
}
\description{
Use copulae to transfer every omics layer to the normal realm
}
\examples{
# To apply the copula-based transition to the normal realm, it is sufficient
# to provide the input data as a named R list, with each element being the
# data set of an omics layer.
copulized <- copulize(multi_omics_micro, verbose = FALSE)

}
