% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\encoding{UTF-8}
\name{get_network}
\alias{get_network}
\title{Extract a \code{carmon} network}
\usage{
get_network(carmon_obj, labels = NULL)
}
\arguments{
\item{carmon_obj}{The object of \code{S3} class \code{carmon} or of \code{S3}
class \code{carmon_rec}.}

\item{labels}{Optional. Used only when \code{carmon_obj} is of the \code{S3} class
\code{carmon_rec}, a vector of strings containing all the variable names.}
}
\value{
\code{get_network()} returns the selected network, in the form of an
object of class \code{igraph}. If using an object of class \code{carmon_rec}, nodes
will have no labels, unless provided through \code{labels}.
}
\description{
\code{get_network()} extracts the reconstructed (and selected) network from a
\code{carmon} object.
}
\examples{
data(multi_omics_small)
c_obj <- carmon(multi_omics_small,
    net_method = "correlation",
    cor_quant = 0.25, analysis = FALSE, plot = FALSE,
    verbose = FALSE
)
network <- get_network(c_obj)

}
