% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.carmon}
\alias{plot.carmon}
\title{Plot \code{carmon} networks}
\usage{
\method{plot}{carmon}(x, node_labels = TRUE, hide_isolated = TRUE, hot_nodes = TRUE, ...)
}
\arguments{
\item{x}{The object of \code{S3} class \code{carmon}.}

\item{node_labels}{Show node names in the network. Defaults to TRUE.}

\item{hide_isolated}{Hide nodes that are not connected to any other node.
Defaults to TRUE.}

\item{hot_nodes}{Highlight nodes that are selected as key nodes by the
node centrality analysis, by drawing them bigger and adding the criteria
for which they result to be central. Defaults to TRUE.}

\item{...}{System required, not used here.}
}
\value{
Returns NULL, invisibly.
}
\description{
\code{plot.carmon()} plots an annotated network from an object of \code{S3} class
\code{carmon}. Nodes from different omics layers will have different color coding.
To plot the network, it's enough to use \code{plot()} call on the \code{carmon} object.
}
\examples{

\donttest{
c_obj <- carmon(multi_omics_small, net_method = "correlation",
                cor_quant = 0.25, plot = FALSE, verbose = FALSE)
plot(c_obj)
}

}
