% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_report}
\alias{plot_report}
\title{Plot the results of the node centrality analysis}
\usage{
plot_report(carmon_obj, scaled = TRUE)
}
\arguments{
\item{carmon_obj}{The object of \code{S3} class \code{carmon}.}

\item{scaled}{Whether the centrality measures reported by the centrality
analysis are 0-1 scaled or not. Defaults to TRUE.}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot the results of the node centrality analysis
}
\examples{
\donttest{
# Let's build and analyse a carmon network:
c_obj <- carmon(multi_omics_small, net_method = "correlation",
                cor_quant = 0.25, analysis = TRUE, plot = FALSE,
                # analysis is already TRUE by default
                verbose = FALSE)
# To plot the results of the centrality analysis:
plot_report(c_obj)
}
}
