% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which_marginals}
\alias{which_marginals}
\title{Print the marginal distributions currently implemented inside carmon}
\usage{
which_marginals()
}
\value{
\code{which_marginals()} prints a message with the marginal distributions
implemented inside carmon in the current version of the package, together
with the nomenclature and the abbreviations that carmon can comprehend when
giving a custom input to the argument \code{marginals}.
}
\description{
Print the marginal distributions currently implemented inside carmon
}
\examples{
# See the marginal distributions you can select to model your omics data with
# carmon
which_marginals()
}
