% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which_omics}
\alias{which_omics}
\title{Print the omics types for which carmon is tailored on}
\usage{
which_omics()
}
\value{
\code{which_omics()} prints a message with the omics types that carmon is
tailored on, together with a list of the synonyms of each type that carmon
can comprehend, and the default tailored marginal distribution of each
omics type.
}
\description{
Print the omics types for which carmon is tailored on
}
\examples{
# See the compatible omics types, their synonyms, and their distributions
which_omics()
}
