% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carmon-package.R
\docType{package}
\name{carmon-package}
\alias{carmon-package}
\title{carmon: Copula-Aided Reconstruction of Multi-Omics Networks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reconstructs networks from multi-omics data sets, harnessing the original probability distribution of the omics layers, through the use of copulas. Use the main function carmon() to perform all the functional steps of the package. These encompass copula-based data transformation, network reconstrution, network analysis, and plot of the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrQuestion/carmon}
  \item \url{https://drquestion.github.io/carmon/}
  \item Report bugs at \url{https://github.com/DrQuestion/carmon/issues}
}

}
\author{
\strong{Maintainer}: Alessio Albanese \email{alessio.albanese.bioinf@gmail.com} (\href{https://orcid.org/0000-0003-1783-5613}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Pariya Behrouzi \email{pariya.behrouzi@wur.nl} (\href{https://orcid.org/0000-0001-6762-5433}{ORCID})
}

}
\keyword{internal}
